+++ cBugReport.py
@@ -235,21 +235,18 @@ class cBugReport(object):
         "sName": "Application and cdb output log",
         "sContent": sCdbStdIOHTML
       });
-      dsAdditionalSummaryInformation = {};
-      if oBugReport.oCdbWrapper.asApplicationCommandLine:
-        dsAdditionalSummaryInformation["Command line"] = " ".join(oBugReport.oCdbWrapper.asApplicationCommandLine);
       # Stick everything together.
       oBugReport.sDetailsHTML = sDetailsHTMLTemplate % {
         "sId": oCdbWrapper.fsHTMLEncode(oBugReport.sId),
-        "sBugDescription": oCdbWrapper.fsHTMLEncode(oBugReport.sBugDescription),
         "sBugLocation": oCdbWrapper.fsHTMLEncode(oBugReport.sBugLocation),
+        "sBugDescription": oCdbWrapper.fsHTMLEncode(oBugReport.sBugDescription),
+        "sBinaryVersion": sBinaryVersionHTML,
+        "sOptionalSource": oBugReport.sBugSourceLocation and \
+            "<tr><td>Source: </td><td>%s</td></tr>" % oBugReport.sBugSourceLocation or "",
         "sSecurityImpact": oBugReport.sSecurityImpact and \
               '<span class="SecurityImpact">%s</span>' % oCdbWrapper.fsHTMLEncode(oBugReport.sSecurityImpact) or "Denial of Service",
-        "sBinaryVersion": sBinaryVersionHTML,
-        "sAdditionalSummaryInformation": "\r\n".join([
-          "          <tr><td>%s:   &nbsp;</td><td>%s</td></tr>" % (oCdbWrapper.fsHTMLEncode(sName).replace(" ", "&nbsp;"), oCdbWrapper.fsHTMLEncode(sValue))
-          for (sName, sValue) in dsAdditionalSummaryInformation.items()
-        ]),
+        "sOptionalCommandLine": oBugReport.oCdbWrapper.asApplicationCommandLine and \
+            "<tr><td>Command line: </td><td>%s</td></tr>" % oBugReport.oCdbWrapper.asApplicationCommandLine or "",
         "sBlocks": "".join(asBlocksHTML),
         "sCdbStdIO": sCdbStdIOHTML,
         "sBugIdVersion": sVersion,