+++ sDetailsHTMLTemplate.py
@@ -65,6 +65,9 @@ sDetailsHTMLTemplate = ("""
       h2 {
         font-size: 120%%;
       }
+      td:first-child {
+        white-space: pre;
+      }
       td {
         vertical-align: top;
       }
@@ -214,12 +217,13 @@ sDetailsHTMLTemplate = ("""
       <h1 class="BlockHeader">BugId %(sId)s @ %(sBugLocation)s summary<span class="BlockHeaderIcon"></span></h1>
       <div class="Content">
         <table>
-          <tr><td>BugId:                  &nbsp;</td><td><span class="Important"><b>%(sId)s</b></span></td></tr>
-          <tr><td>Location:               &nbsp;</td><td><span class="Important">%(sBugLocation)s</span></td></tr>
-          <tr><td>Description:            &nbsp;</td><td><span class="Important">%(sBugDescription)s</span></td></tr>
-          <tr><td>Version:                &nbsp;</td><td>%(sBinaryVersion)s</td></tr>
-          <tr><td>Security&nbsp;impact:   &nbsp;</td><td>%(sSecurityImpact)s</td></tr>
-%(sAdditionalSummaryInformation)s
+          <tr><td>BugId:           </td><td><span class="Important"><b>%(sId)s</b></span></td></tr>
+          <tr><td>Location:        </td><td><span class="Important">%(sBugLocation)s</span></td></tr>
+          <tr><td>Description:     </td><td><span class="Important">%(sBugDescription)s</span></td></tr>
+          <tr><td>Version:         </td><td>%(sBinaryVersion)s</td></tr>
+%(sOptionalSource)s
+          <tr><td>Security impact: </td><td>%(sSecurityImpact)s</td></tr>
+%(sOptionalCommandLine)s
         </table>
         <br/>
         BugId version <b>%(sBugIdVersion)s</b>. You may not use this version of BugId for commercial purposes. Please